+++ client/types/index.d.ts
@@ -75,6 +75,7 @@ declare global {
     regularity: string
     tagDTO: TagDTO
     frequencyProperty: FrequencyPropertyDTO
+    bookletIds?: string[]
   }
 
   interface FrequencyPropertyDTO {
@@ -108,6 +109,10 @@ declare global {
     frequencyProperty: FrequencyPropertyDTO
     bookletIds: string[]
     recurrenceRule: RecurrenceRuleDTO
+  }
+
+  interface RegularTransactionsDeletionRequest {
+    transactionIds: string[]
   }
   interface BookletReport {
     label: string